requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
requests==2.32.3
python-dotenv==1.1.0
pypdf==5.4.0
PyYAML==6.0.2
